rename folder (mark as "__OLD")
