+++ code/system.ui/react.monaco/src/index.mts
@@ -8,7 +8,7 @@ export { Pkg };
  * Code Editor
  */
 export { MonacoEditor } from './ui/ui.MonacoEditor';
-export { MonacoCrdt } from './ui.logic/MonacoCrdt';
+export { MonacoCrdt } from './ui.logic/MonacoCrdt__OLD';
 
 /**
  * Dev
