+++ code/system.ui/react.monaco/src/test.ui/entry.Specs.mts
@@ -2,7 +2,7 @@ export { Pkg } from '../index.pkg.mjs';
 
 export const Specs = {
   'sys.ui.code.ui.MonacoEditor': () => import('../ui/ui.MonacoEditor/-SPEC'),
-  'sys.ui.code.ui.MonacoCrdt': () => import('../ui.logic/MonacoCrdt/-dev/-SPEC'),
+  'sys.ui.code.ui.MonacoCrdt': () => import('../ui.logic/MonacoCrdt__OLD/-dev/-SPEC'),
   'sys.ui.code.ui.Sample.01': () => import('../ui/ui.Sample.01/-SPEC'),
 };
