+++ code/system.ui/react.monaco/src/ui.logic/MonacoCrdt__OLD/common.ts
