+++ code/system.ui/react.monaco/src/ui.logic/index.ts
@@ -1,2 +1,2 @@
 export * from './Editor.Carets';
-export * from './MonacoCrdt';
+export * from './MonacoCrdt__OLD';
